requirements.txt
asttokens==2.2.1
azure-appconfiguration==1.4.0
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
docker==6.1.2
dominate==2.8.0
executing==1.2.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
greenlet==2.0.2
idna==3.4
ipykernel==6.23.1
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msal==1.22.0
msal-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
portalocker==2.7.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
sodapy==2.2.0
SQLAlchemy==2.0.15
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
visitor==0.1.3
wcwidth==0.2.6
websocket-client==1.5.2
Werkzeug==2.3.4
